# frontend/src/lib/api/schema/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlametteMoney.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get user settings",
        "description": "Returns current user settings.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update user settings",
        "description": "Updates current user settings.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/portfolio-balance-series": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get portfolio balance series",
        "description": "Returns portfolio balance time-series converted to a selected base currency.",
        "parameters": [
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Interval",
            "in": "query",
            "schema": {
              "default": "Auto",
              "$ref": "#/components/schemas/ReportInterval"
            }
          },
          {
            "name": "BaseCurrency",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AccountIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioBalanceSeriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PortfolioAccountResponse": {
        "required": [
          "id",
          "name",
          "color",
          "currency",
          "initialBalance",
          "currentBalance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "initialBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currentBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "PortfolioBalancePointResponse": {
        "required": [
          "bucketKey",
          "bucketLabel",
          "bucketDate",
          "totalBalance",
          "accountBalances",
          "totalsByCurrency",
          "missingCurrencies"
        ],
        "type": "object",
        "properties": {
          "bucketKey": {
            "type": "string"
          },
          "bucketLabel": {
            "type": "string"
          },
          "bucketDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "accountBalances": {
            "type": "object",
            "additionalProperties": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          "totalsByCurrency": {
            "type": "object",
            "additionalProperties": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          "missingCurrencies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PortfolioBalanceSeriesResponse": {
        "required": [
          "baseCurrency",
          "startDate",
          "endDate",
          "interval",
          "accounts",
          "points",
          "summary"
        ],
        "type": "object",
        "properties": {
          "baseCurrency": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "interval": {
            "$ref": "#/components/schemas/ReportInterval"
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioAccountResponse"
            }
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioBalancePointResponse"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/PortfolioBalanceSummaryResponse"
          }
        }
      },
      "PortfolioBalanceSummaryResponse": {
        "required": [
          "startBalance",
          "endBalance",
          "delta",
          "deltaPercent",
          "pointCount",
          "dayCount"
        ],
        "type": "object",
        "properties": {
          "startBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "endBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "delta": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "deltaPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "pointCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "dayCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ReportInterval": {
        "enum": [
          "Auto",
          "None",
          "Day",
          "Week",
          "Month"
        ]
      },
...
      "UpdateUserSettingsRequest": {
        "required": [
          "baseCurrency"
        ],
        "type": "object",
        "properties": {
          "baseCurrency": {
            "type": "string"
          }
        }
      },
      "UserSettingsResponse": {
        "required": [
          "baseCurrency"
        ],
        "type": "object",
        "properties": {
          "baseCurrency": {
            "type": "string"
          }
        }
      }
    }
  },
...
}